+++ src/components/TeacherComponents/UnitList/UnitList.tsx
@@ -159,8 +159,21 @@ const UnitList: FC<UnitListProps> = (props) => {
     category === "reading-writing-oracy"
       ? "Reading, writing & oracy"
       : category;
-  const newPageItems = getPageItems(currentPageItems, false);
-  const legacyPageItems = getPageItems(currentPageItems, true);
+  const newPageItems = getPageItems({
+    pageItems: currentPageItems,
+    pickLegacyItems: false,
+    isSwimming: false,
+  });
+  const legacyPageItems = getPageItems({
+    pageItems: currentPageItems,
+    pickLegacyItems: true,
+    isSwimming: false,
+  });
+  const swimmingPageItems = getPageItems({
+    pageItems: currentPageItems,
+    pickLegacyItems: false,
+    isSwimming: true,
+  });
 
   const { phaseSlug, keyStageSlug, examBoardSlug } = getProgrammeFactors(props);
   const indexOfFirstLegacyUnit = units
@@ -326,12 +339,42 @@ const UnitList: FC<UnitListProps> = (props) => {
       />
     ) : null;
 
+  const SwimmingUnits = () => {
+    if (swimmingPageItems.length && keyStageSlug && phaseSlug) {
+      const title = `${swimmingPageItems[0]?.[0]?.groupUnitsAs} units (all years)`;
+      return (
+        <OakUnitsContainer
+          isLegacy={false}
+          subject={""}
+          phase={phaseSlug}
+          curriculumHref={resolveOakHref({
+            page: "curriculum-previous-downloads",
+            query: {
+              subject: linkSubject,
+              keystage: keyStageSlug,
+            },
+          })}
+          showHeader={
+            swimmingPageItems.length || indexOfFirstLegacyUnit % pageSize === 0
+              ? true
+              : false
+          }
+          unitCards={getUnitCards(swimmingPageItems)}
+          isCustomUnit={true}
+          customHeadingText={title}
+          bannerText="Swimming and water safety lessons should be selected based on the ability and experience of your pupils."
+        />
+      );
+    } else return null;
+  };
+
   return (
     <OakFlex $flexDirection="column">
       <OakAnchorTarget id="unit-list" />
       {currentPageItems.length ? (
         isUnitListData(props) ? (
           <OakFlex $flexDirection="column" $gap="space-between-xxl">
+            <SwimmingUnits />
             <NewUnits category={modifiedCategory} />
             <LegacyUnits />
           </OakFlex>