+++ src/components/TeacherComponents/UnitList/helpers.tsx
@@ -22,17 +22,37 @@ export const isUnitListData = (
   return (u as UnitListingData).keyStageSlug !== undefined;
 };
 
-export const getPageItems = (
-  pageItems: CurrentPageItemsProps[] | SpecialistUnit[][],
-  pickLegacyItems: boolean,
-) => {
-  if (!isCurrentPageItems(pageItems)) {
-    return [];
+export const getPageItems = ({
+  pageItems,
+  pickLegacyItems,
+  isSwimming,
+}: {
+  pageItems: CurrentPageItemsProps[] | SpecialistUnit[][];
+  pickLegacyItems: boolean;
+  isSwimming: boolean;
+}) => {
+  switch (true) {
+    case !isCurrentPageItems(pageItems):
+      return [];
+    case isSwimming:
+      return pageItems.filter((item) => {
+        const isSwimmingItem =
+          item[0]!.groupUnitsAs === "Swimming and water safety";
+        return isSwimmingItem;
+      });
+    case pickLegacyItems:
+      return pageItems.filter((item) => {
+        const legacyItem = isSlugLegacy(item[0]!.programmeSlug);
+        return legacyItem;
+      });
+    default:
+      return pageItems.filter((item) => {
+        const legacyItem = isSlugLegacy(item[0]!.programmeSlug);
+        const groupUnitsAsItem = item[0]!.groupUnitsAs;
+        const filteredItem = !legacyItem && !groupUnitsAsItem;
+        return filteredItem;
+      });
   }
-  return pageItems.filter((item) => {
-    const isLegacy = isSlugLegacy(item[0]!.programmeSlug);
-    return pickLegacyItems ? isLegacy : !isLegacy;
-  });
 };
 
 export const getProgrammeFactors = (props: UnitListProps) => {